# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Validation Center API
  version: 0.1.0
  description: >
    Phase 0 + Phase 1 contract for the AI-assisted Data Comparison /
    Reconciliation Platform (deterministic core). AI suggestion (P4) and PDF
    extraction (P3) endpoints are reserved and not included here.
    Principle: AI suggests, humans approve, the rule engine judges.
servers:
  - url: /api/v1

# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
...
paths:
...
  /templates:
    get:
      tags: [Templates]
      summary: List templates
      parameters:
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Template' }
    post:
      tags: [Templates]
      summary: Create a template (creates an initial DRAFT version)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TemplateCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Template' }
...
  /datasets:
    post:
      tags: [Datasets]
      summary: Upload a Source/Destination dataset (CSV/Excel). Async profiling.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                role: { $ref: '#/components/schemas/DatasetRole' }
                name: { type: string }
                file: { type: string, format: binary }
      responses:
        '202':
          description: Accepted (profiling queued)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dataset' }
  /datasets/{datasetId}/schema:
    get:
      tags: [Datasets]
      summary: Get inferred columns + profile
      parameters:
        - { name: datasetId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DatasetColumn' }

  # ---------------------------- MAPPINGS ---------------------------------
...
  /recon/runs:
    post:
      tags: [Reconciliation]
      summary: Trigger a reconciliation run (async batch). Requires OPERATOR role.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunCreate' }
      responses:
        '202':
          description: Accepted (queued)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '422': { description: Inputs not in runnable state }
    get:
      tags: [Reconciliation]
      summary: List runs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Run' }
  /recon/runs/{runId}:
    get:
      tags: [Reconciliation]
      summary: Run status + summary
      parameters:
        - { name: runId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunDetail' }
  /recon/runs/{runId}/results:
    get:
      tags: [Reports]
      summary: Drill-down rows (filter by category/field)
      parameters:
        - { name: runId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: category, in: query, schema: { $ref: '#/components/schemas/ResultCategory' } }
        - { name: fieldKey, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 50 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ResultRow' }
  /recon/runs/{runId}/export:
    get:
      tags: [Reports]
      summary: Export report (Excel). Embeds run + version metadata.
      parameters:
        - { name: runId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: format, in: query, schema: { type: string, enum: [xlsx, csv], default: xlsx } }
      responses:
        '200':
          description: Binary file
          content:
            application/octet-stream:
              schema: { type: string, format: binary }

  # ------------------------------ AUDIT ----------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Template:
      type: object
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        name: { type: string }
        description: { type: string }
        currentVersionId: { type: string, format: uuid, nullable: true }
    TemplateCreate:
      type: object
      required: [key, name]
      properties:
        key: { type: string }
        name: { type: string }
        description: { type: string }
...
    Dataset:
      type: object
      properties:
        id: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/DatasetRole' }
        name: { type: string }
        origin: { type: string, enum: [CSV, EXCEL] }
        rowCount: { type: integer, nullable: true }
        status: { type: string, enum: [UPLOADED, PROFILED, READY, FAILED] }
    DatasetColumn:
      type: object
      properties:
        id: { type: string, format: uuid }
        columnName: { type: string }
        ordinal: { type: integer }
        inferredDatatype: { $ref: '#/components/schemas/Datatype' }
        nullCount: { type: integer, nullable: true }
        distinctCount: { type: integer, nullable: true }
        sampleValues: { type: array, items: { type: string } }
...
    RunCreate:
      type: object
      required: [templateVersionId, mappingId, sourceDatasetId, destDatasetId]
      properties:
        templateVersionId: { type: string, format: uuid }
        mappingId: { type: string, format: uuid }
        sourceDatasetId: { type: string, format: uuid }
        destDatasetId: { type: string, format: uuid }
    Run:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELLED] }
        ruleEngineVersion: { type: string }
        queuedAt: { type: string, format: date-time }
    RunDetail:
      allOf:
        - $ref: '#/components/schemas/Run'
        - type: object
          properties:
            templateVersionId: { type: string, format: uuid }
            mappingId: { type: string, format: uuid }
            summary: { $ref: '#/components/schemas/Summary' }
    Summary:
      type: object
      properties:
        totalRecords: { type: integer }
        matchCount: { type: integer }
        mismatchCount: { type: integer }
        missingSrc: { type: integer }
        missingDst: { type: integer }
        duplicateCount: { type: integer }
        exceptionCount: { type: integer }
        bySeverity: { type: object, additionalProperties: { type: integer } }
        byField: { type: object, additionalProperties: { type: integer } }
    ResultRow:
      type: object
      properties:
        recordKey: { type: string }
        category: { $ref: '#/components/schemas/ResultCategory' }
        fieldKey: { type: string, nullable: true }
        ruleId: { type: string, format: uuid, nullable: true }
        severity: { type: string, enum: [ERROR, WARNING, INFO], nullable: true }
        expected: { type: string, nullable: true }
        actual: { type: string, nullable: true }
        verdict: { type: string, nullable: true }
        detail: { type: object, additionalProperties: true }
...
    Datatype:
      type: string
      enum: [STRING, INTEGER, DECIMAL, DATE, DATETIME, BOOLEAN, ENUM]
...
    DatasetRole:
      type: string
      enum: [SOURCE, DESTINATION]
    ResultCategory:
      type: string
      enum: [MATCH, MISMATCH, MISSING_IN_SOURCE, MISSING_IN_DEST, DUPLICATE, EXCEPTION]